Move ProjectSettings to its own file
