+++ packages/api-docs-builder-core/.eslintrc.js
@@ -1,0 +1,6 @@
+module.exports = {
+  rules: {
+    'import/no-default-export': 'error',
+    'import/prefer-default-export': 'off',
+  },
+};
