+++ packages/api-docs-builder/.eslintrc.js
@@ -1,0 +1,5 @@
+module.exports = {
+  rules: {
+    'import/prefer-default-export': 'off',
+  },
+};
