+++ packages/api-docs-builder/index.ts
@@ -1,0 +1,2 @@
+export { buildApi } from './buildApi';
+export type { ProjectSettings } from './ProjectSettings';
