+++ packages/api-docs-builder/package.json
@@ -2,7 +2,7 @@
   "name": "@mui-internal/api-docs-builder",
   "version": "1.0.0",
   "private": "true",
-  "main": "./buildApi.ts",
+  "main": "./index.ts",
   "scripts": {
     "test": "cd ../../ && cross-env NODE_ENV=test mocha 'packages/api-docs-builder/**/*.test.{js,ts,tsx}'",
     "typescript": "tsc -p tsconfig.json"
